+++ scout/server/blueprints/variants/views.py
@@ -173,7 +173,6 @@ def str_variants(institute_id, case_name):
 
     # populate filters dropdown
     available_filters = list(store.filters(institute_id, category))
-
     form.filters.choices = [
         (filter.get("_id"), filter.get("display_name")) for filter in available_filters
     ]
@@ -248,12 +247,21 @@ def sv_variants(institute_id, case_name):
 
     # update status of case if visited for the first time
     controllers.activate_case(store, institute_obj, case_obj, current_user)
+
     form = controllers.populate_sv_filters_form(store, institute_obj, case_obj, category, request)
+
+    # populate filters dropdown
+    available_filters = store.filters(institute_obj["_id"], category)
+    form.filters.choices = [
+        (filter.get("_id"), filter.get("display_name")) for filter in available_filters
+    ]
 
     # Populate chromosome select choices
     controllers.populate_chrom_choices(form, case_obj)
 
     cytobands = store.cytoband_by_chrom(case_obj.get("genome_build"))
+
+    form = controllers.update_form_hgnc_symbols(store, case_obj, form)
 
     variants_query = store.variants(case_obj["_id"], category=category, query=form.data)
 
@@ -273,6 +281,7 @@ def sv_variants(institute_id, case_name):
         dismiss_variant_options=DISMISS_VARIANT_OPTIONS,
         variant_type=variant_type,
         form=form,
+        filters=available_filters,
         cytobands=cytobands,
         severe_so_terms=SEVERE_SO_TERMS,
         manual_rank_options=MANUAL_RANK_OPTIONS,